Product Admin - secure the views

+++ products/views.py
@@ -2,6 +2,7 @@ from django.shortcuts import render, redirect, reverse, get_object_or_404
 from django.contrib import messages
 from django.db.models import Q
 from django.db.models.functions import Lower
+from django.contrib.auth.decorators import login_required
 
 from .models import Product, Category
 from .forms import ProductForm
@@ -88,10 +89,16 @@ def product_detail(request, product_id):
 
 
 # Add Product view
+@login_required
 def add_product(request):
     """
     Add a product to the store
     """
+    # Restrict access to superusers only
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only store owners can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         # Instantiate the submitted form
         form = ProductForm(request.POST, request.FILES)
@@ -115,8 +122,14 @@ def add_product(request):
 
 
 # Edit product view
+@login_required
 def edit_product(request, product_id):
     """ Edit a product in the store """
+    # Restrict access to superusers only
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES, instance=product)
@@ -140,10 +153,16 @@ def edit_product(request, product_id):
 
 
 # Delete Product view
+@login_required
 def delete_product(request, product_id):
     """
     Delete a product from the store
     """
+    # Restrict access to superusers only
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry only store owners can do that.')
+        return redirect(reverse('home'))
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted')
